requirements.txt
anyio==4.13.0
asttokens==3.0.1
certifi==2026.2.25
charset-normalizer==3.4.7
colorama==0.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
loguru==0.7.3
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
packaging==26.0
parso==0.8.6
platformdirs==4.9.4
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyzmq==27.1.0
requests==2.33.1
six==1.17.0
stack-data==0.6.3
tornado==6.5.5
traitlets==5.14.3
urllib3==2.6.3
wcwidth==0.6.0
win32_setctime==1.2.0
